+++ src/components/Sidebar/index.tsx
@@ -1,7 +1,7 @@
-import * as Dialog from '@radix-ui/react-dialog'
-import NewPageModal from '../Modal/NewPageModal'
-import NavList from '../NavList'
 import Menu from '../Menu'
+import MenuButton from '../MenuButton'
+import NavList from '../NavList'
+import NewPageModal from '../Modal/NewPageModal'
 
 import IconPlusRounded from '@/assets/icons/IconPlusRounded'
 
@@ -12,18 +12,9 @@ const Sidebar = () => {
 
       <NavList />
 
-      <Dialog.Root>
-        <Dialog.Trigger className="rounded-md focus:outline-none focus-visible:ring-1.5 dark:focus-visible:ring-gray-300 focus-visible:ring-white-300">
-          <div className="flex flex-row items-center gap-2 h-9 px-2 rounded-md dark:hover:bg-gray-700 hover:bg-white-700/75 dark:active:bg-gray-800/80 active:bg-white-600/75">
-            <div className="flex justify-center items-center w-6">
-              <IconPlusRounded width="17" height="17" />
-            </div>
-            <span className="flex items-center font-int text-sm dark:text-gray-100 text-gray-300">Nova página</span>
-          </div>
-        </Dialog.Trigger>
-
+      <MenuButton icon={<IconPlusRounded width="17" height="17" />} label="Nova página" triggerType="dialog">
         <NewPageModal />
-      </Dialog.Root>
+      </MenuButton>
     </div>
   )
 }
