+++ src/types/themeTypes.ts
@@ -1,7 +1,9 @@
+import { ReactNode } from 'react'
+
 export type Theme = 'light' | 'dark' | 'system'
 
 export interface ThemeProviderProps {
-  children: React.ReactNode
+  children: ReactNode
 }
 
 export interface ThemeContextType {
